+++ tests/test_on_data/ophys/test_miniscope_converter.py
@@ -146,8 +146,8 @@ class TestMiniscopeConverterLegacyTyeLabFormat:
         self.converter = MiniscopeConverter(folder_path=self.folder_path)
         self.test_dir = Path(tempfile.mkdtemp())
 
-        self.stub_frames = 2
-        self.conversion_options = dict(stub_test=True, stub_frames=self.stub_frames)
+        self.stub_samples = 2
+        self.conversion_options = dict(stub_test=True, stub_samples=self.stub_samples)
 
         self.device_name = "Miniscope"
         self.device_metadata = dict(
@@ -220,7 +220,7 @@ class TestMiniscopeConverterLegacyTyeLabFormat:
             nwbfile = io.read()
 
         num_frames = nwbfile.acquisition[self.photon_series_name].data.shape[0]
-        assert num_frames == self.stub_frames
+        assert num_frames == self.stub_samples
 
     def test_run_conversion_updated_metadata(self):
         """Test conversion with updated metadata."""
@@ -280,7 +280,7 @@ class TestMiniscopeConverterLegacyTyeLabFormat:
             nwbfile = io.read()
 
         num_frames = nwbfile.acquisition[self.photon_series_name].data.shape[0]
-        assert num_frames == self.stub_frames
+        assert num_frames == self.stub_samples
 
     def test_converter_in_converter_pipe(self):
         """Test MiniscopeConverter in ConverterPipe."""
@@ -302,4 +302,4 @@ class TestMiniscopeConverterLegacyTyeLabFormat:
         with NWBHDF5IO(path=nwbfile_path) as io:
             nwbfile = io.read()
         num_frames = nwbfile.acquisition[self.photon_series_name].data.shape[0]
-        assert num_frames == self.stub_frames
+        assert num_frames == self.stub_samples